micromath/fibonacci/requirements.txt
Flask~=1.1.2
gunicorn~=20.0.4
jsonschema~=3.2.0
python-logstash-async~=2.1.1
